+++ src/apps/apps_factory.rs
@@ -26,7 +26,7 @@ use super::http::{Caddy, Curlie, Hurl, Restish, Xh};
 use super::logs::{Gonzo, Hl, LazyJournal, Lnav, Logdy, Loggo, Nerdlog, Rhit, Tailspin, Vector};
 use super::networking::{Boring, Dog, Doggo};
 use super::shell::{Atuin, Axe, Carapace, Fzf, Rgxg, RustParallel, Skim, Starship, Vivid, Zoxide};
-use super::system::{Bottom, Btop, Duf, Dust, Dysk, Erdtree, Netwatch, Procs, Syswatch};
+use super::system::{Bottom, Btop, Diskwatch, Duf, Dust, Dysk, Erdtree, Netwatch, Procs, Syswatch};
 
 pub fn create_app(
     id: &str, gh_token: Option<String>, cb_token: Option<String>, _gl_token: Option<String>,
@@ -61,6 +61,7 @@ pub fn create_app(
         DockMate::ID => Some(Box::new(DockMate::new(client))),
         Doppler::ID => Some(Box::new(Doppler::new(client))),
         Dry::ID => Some(Box::new(Dry::new(client))),
+        Diskwatch::ID => Some(Box::new(Diskwatch::new(client))),
         Duf::ID => Some(Box::new(Duf::new(client))),
         Erdtree::ID => Some(Box::new(Erdtree::new(client))),
         Dust::ID => Some(Box::new(Dust::new(client))),